Removal of v1.x only commands
