+++ ublox-bluetooth-commands_test.go
@@ -22,32 +22,75 @@ func TestUbloxBluetoothCommands(t *testing.T) {
 	defer ub.Close()
 
 	serial.SetVerbose(true)
-	for i := 0; i < 1; i++ {
-		exerciseTheDevice("D5926479C652r", ub, t, i, true, false)
-	}
-}
 
-func TestAbort(t *testing.T) {
-	ub, err := setupBluetooth()
-	if err != nil {
-		t.Fatalf("setupBluetooth error %v\n", err)
-	}
-	defer ub.Close()
-
-	// put some events into the sensors logs (assuming connection events are being logged)
-	for i := 0; i < 20; i++ {
-		fmt.Printf("Connect attempt %d\n", i)
-		errr := connectToDevice("D5926479C652r", func(t *testing.T) error {
-			return ub.DisconnectFromDevice()
-		}, ub, t)
-		if errr != nil {
-			err = errors.Wrapf(err, "connectToDevice error: %v", errr)
+	err = connectToDevice("EE9EF8BA058Br", func(t *testing.T) error {
+		fmt.Printf("[GetTime] starting\n")
+		time, err := ub.GetTime()
+		if err != nil {
+			t.Errorf("GetTime error %v\n", err)
 		}
-	}
+		fmt.Printf("[GetTime] Current timestamp %d\n", time)
+
+		version, err := ub.GetVersion()
+		if err != nil {
+			t.Fatalf("GetVersion error %v\n", err)
+		}
+		fmt.Printf("[GetVersion] replied with: %v\n", version)
+
+		if version.SoftwareVersion != "2.0" {
+			t.Fatalf("Cannot continue with version %s, needs to be 2.0\n", version.SoftwareVersion)
+		}
+
+		config, err := ub.ReadConfig()
+		if err != nil {
+			t.Fatalf("ReadConfig error %v\n", err)
+		}
+		fmt.Printf("[ReadConfig] replied with: %v\n", config)
+
+		/*if downloadEvents {
+			startingIndex := info.CurrentSequenceNumber - info.RecordsCount
+			fmt.Printf("[DownloadLogFile] starting run: %d\n", itteration)
+			err = ub.DownloadLogFile(startingIndex, func(b []byte) error {
+				//fmt.Print(".")
+				return nil
+			})
+			if err != nil {
+				t.Fatalf("DownloadLogFile error %v\n", err)
+			}
+			fmt.Printf("[DownloadLogFile] complete\n")
+		}
+
+		if downloadSlotData {
+			slotCount, err := ub.ReadSlotCount()
+			if err != nil {
+				t.Errorf("ReadSlotCount error %v\n", err)
+			} else {
+				fmt.Printf("[ReadSlotCount] replied with: %v\n", slotCount)
+				slotInfo, err := ub.ReadSlotInfo(0)
+				if err != nil {
+					t.Errorf("ReadSlotInfo error %v\n", err)
+				} else {
+					fmt.Printf("[ReadSlotInfo] replied with: %v\n", slotInfo)
+					slotData, err := ub.ReadSlotData(0, 0, slotInfo.Bytes)
+					if err != nil {
+						t.Errorf("ReadSlotData error %v\n", err)
+					}
+					fmt.Printf("[ReadSlotData] replied with: %v\n", slotData)
+				}
+			}
+		}*/
+
+		err = ub.DisconnectFromDevice()
+		if err != nil {
+			t.Errorf("DisconnectFromDevice error %v\n", err)
+		}
+		return err
+	}, ub, t)
 
 	if err != nil {
-		t.Fatalf("TestAbort error %v\n", err)
+		t.Errorf("exerciseTheDevice error %v\n", err)
 	}
+
 }
 
 func TestPagedDownloads(t *testing.T) {
@@ -57,8 +100,7 @@ func TestPagedDownloads(t *testing.T) {
 	}
 	defer ub.Close()
 
-	// C1851F6083F8r or CE1A0B7E9D79r
-	err = connectToDevice("D5926479C652r", func(t *testing.T) error {
+	err = connectToDevice("EE9EF8BA058Br", func(t *testing.T) error {
 		defer ub.DisconnectFromDevice()
 
 		err := ub.EnableNotifications()
@@ -87,7 +129,7 @@ func TestEventClear(t *testing.T) {
 	}
 	defer ub.Close()
 
-	err = connectToDevice("CE1A0B7E9D79r", func(t *testing.T) error {
+	err = connectToDevice("EE9EF8BA058Br", func(t *testing.T) error {
 		defer ub.DisconnectFromDevice()
 
 		t.Fatalf("TODO - rework")
@@ -117,7 +159,7 @@ func TestSlotDataClear(t *testing.T) {
 	}
 	defer ub.Close()
 
-	err = connectToDevice("F344B0C992E1r", func(t *testing.T) error {
+	err = connectToDevice("EE9EF8BA058Br", func(t *testing.T) error {
 		defer ub.DisconnectFromDevice()
 
 		startingCount, err := ub.ReadSlotCount()
@@ -152,7 +194,7 @@ func TestReboot(t *testing.T) {
 	}
 	defer ub.Close()
 
-	err = connectToDevice("CE1A0B7E9D79r", func(t *testing.T) error {
+	err = connectToDevice("EE9EF8BA058Br", func(t *testing.T) error {
 		defer ub.DisconnectFromDevice()
 		ub.PeerList()
 		return nil
@@ -229,70 +271,4 @@ func connectToDevice(mac string, tfn TestFunc, ub *UbloxBluetooth, t *testing.T)
 		fmt.Println("Disconnected!")
 		return nil
 	})
-}
-
-func exerciseTheDevice(deviceAddr string, ub *UbloxBluetooth, t *testing.T, itteration int, downloadEvents bool, downloadSlotData bool) {
-	err := connectToDevice(deviceAddr, func(t *testing.T) error {
-		fmt.Printf("[GetTime] starting\n")
-		time, err := ub.GetTime()
-		if err != nil {
-			t.Errorf("GetTime error %v\n", err)
-		}
-		fmt.Printf("[GetTime] Current timestamp %d\n", time)
-
-		version, err := ub.GetVersion()
-		if err != nil {
-			t.Fatalf("GetVersion error %v\n", err)
-		}
-		fmt.Printf("[GetVersion] replied with: %v\n", version)
-
-		config, err := ub.ReadConfig()
-		if err != nil {
-			t.Fatalf("ReadConfig error %v\n", err)
-		}
-		fmt.Printf("[ReadConfig] replied with: %v\n", config)
-
-		/*if downloadEvents {
-			startingIndex := info.CurrentSequenceNumber - info.RecordsCount
-			fmt.Printf("[DownloadLogFile] starting run: %d\n", itteration)
-			err = ub.DownloadLogFile(startingIndex, func(b []byte) error {
-				//fmt.Print(".")
-				return nil
-			})
-			if err != nil {
-				t.Fatalf("DownloadLogFile error %v\n", err)
-			}
-			fmt.Printf("[DownloadLogFile] complete\n")
-		}
-
-		if downloadSlotData {
-			slotCount, err := ub.ReadSlotCount()
-			if err != nil {
-				t.Errorf("ReadSlotCount error %v\n", err)
-			} else {
-				fmt.Printf("[ReadSlotCount] replied with: %v\n", slotCount)
-				slotInfo, err := ub.ReadSlotInfo(0)
-				if err != nil {
-					t.Errorf("ReadSlotInfo error %v\n", err)
-				} else {
-					fmt.Printf("[ReadSlotInfo] replied with: %v\n", slotInfo)
-					slotData, err := ub.ReadSlotData(0, 0, slotInfo.Bytes)
-					if err != nil {
-						t.Errorf("ReadSlotData error %v\n", err)
-					}
-					fmt.Printf("[ReadSlotData] replied with: %v\n", slotData)
-				}
-			}
-		}*/
-
-		err = ub.DisconnectFromDevice()
-		if err != nil {
-			t.Errorf("DisconnectFromDevice Itteration[%d] error %v\n", itteration, err)
-		}
-		return err
-	}, ub, t)
-
-	if err != nil {
-		t.Errorf("exerciseTheDevice error %v\n", err)
-	}
 }
